Add Fix Cost Elements

+++ lib/data_model/new/rag_component_language_model.dart
@@ -26,7 +26,7 @@ class RagComponentLanguageModel {
   }
 
   double calculateCost(
-      bool isInput, LanguageModelPriceComponentTypes type, double amount) {
+      bool isInput, LanguageModelPriceComponentTypes type, int amount) {
     PriceComponent? calculateComponent;
     for (PriceComponent component in priceComponents) {
       if (component.isInput == isInput && component.type == type) {
